Fixed rounding of score in the connection test
There was an inconsistency between scoring rounding in the web/mail test versus connection test.

This breaks the connection test score on Chromium-based browsers (seems like Chromium uses a comma, Firebox uses a dot, not sure if that is why it breaks.

See also: https://github.com/internetstandards/Internet.nl/blob/134609a802ad3dca81cc90cb5f7f19acf898e454/checks/probes.py#L115

+++ interface/views/connection.py
@@ -88,7 +88,7 @@ def results(request, request_id):
         connectionprobes["resolver"].rated_results_by_model(ct),
     ]
     scores = [pr["totalscore"] for pr in probereports]
-    score = max(min(sum(scores) / len(scores), 100), 0)
+    score = max(min(int(sum(scores) / len(scores)), 100), 0)
 
     return render(
         request,
